Base completa del backend de la app

+++ src/main/java/com/gestion/infrastructure/persistence/repositories/UserRepository.java
@@ -7,5 +7,5 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface UserRepository extends JpaRepository<UserDAO, Integer> {
+public interface UserRepository extends JpaRepository<UserDAO, Long> {
 }
